backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot-backend"
version = "1.0.0"
description = "Backend for the RAG Chatbot system"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "sqlalchemy==2.0.23",
    "qdrant-client==1.8.0",
    "cohere==5.5.3",
    "python-dotenv==1.0.0",
    "pyjwt==2.8.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "alembic==1.13.1",
    "uuid7==0.1.0",
    "bleach==6.1.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
ingest-book = "ingestion.content_pipeline:main"